refactor: extract AI spacing page-side machinery into ai-spacing/content-script.ts
classifyCandidates, findCandidates, classifyBatch, the seam type aliases,
and the ambiguous-shape registry sat before the entry points they serve in
content-script.ts, making the entry hard to read. Move that half into
ai-spacing/content-script.ts (exporting classifyBatch), mirroring how
prompt-classifier.ts is the worker half, so the root content-script.ts is
only the entry: the seam assignment, the readyState branch, and the popup
listener. No behavior change.
